+++ boot/espressif/hal/include/mcuboot_config/mcuboot_config.h
@@ -57,8 +57,10 @@
 /* Uncomment to enable the direct-xip code path. */
 /* #define MCUBOOT_DIRECT_XIP */
 
-/* Uncomment to enable the ram-load code path. */
-/* #define MCUBOOT_RAM_LOAD */
+/* Define to enable the ram-load code path. */
+#if defined(CONFIG_BOOT_RAM_LOAD)
+#define MCUBOOT_RAM_LOAD
+#endif
 
 /*
  * Cryptographic settings
